Remove wrongly exposed REMOVE event
POLLREMOVE is actually only used for manipulating
and existing poll set, and has no value in a one-
off `poll` command-line interface.

In a command-line `poll`, the "poll set" is the
string or array of arguments, or whatever data
representation is used to produce it. If you
need to remove a file descriptor between calls
to a command-line `poll`, you would remove it
from that argument list.

We could repurpose REMOVE in a command-line
context to remove FDs specified earlier in
the argument list from the poll set used to
poll, but that seems like really bad design.
Ideally, code should be written so that you
can easily remove FDs from the argument list
before calling the `poll` program at all,
and the simple argument "syntax" of `poll`
makes that reasonably trivial.