Replace inline with static
static is, for all intents and purposes, a
more useful version of inline - a compiler
can inline static functions just as freely,
and unless you want the function to be
extern, there is no benefit to inline that
I can remember.

And in this case static is more appropriate
anyway, because we do not have a proactive
need for the function to be extern, so we
should make it static anyway.