Stop using bool type
It is a nice type, but I am transitioning to strict C89
compatibility. One day I might decide to go back and it
would probably be better for clearer intent-describing,
but for now I am thoroughly used to using integers as
booleans in C, and I still want the back-compat. But I
do think that ideally, if I was less neurotic about
compatibility, I would just use bool and leave it up
to those without the `stdbool.h` header to provide it
themselves.