Stop reusing FD argument for printing
Mixed feelings about this one. In some ways, it
would be better to show the user the string
they passed in. Especially if one day multiple
formats/bases are supported for FD arguments.
In other ways, it is better to show a canonical
representation of the number, without leading
zeroes or whatever else the user puts in. Also,
this program currently combines `struct pollfd`
entries for the same FD, if an FD is given more
than once on the command line, and this way you
always get the canonical FD representation even
if you specified the FD multiple times, each
with slightly different strings (for example,
given the arguments `01 IN 1 HUP` the event
print will always be `1 ...`, never `01 ...`,
which was less guaranteed when both reusing the
agument string for output and doing the merging
of identical FDs, because in a case like `poll
2 OUT 2 HUP 1 OUT 01 HUP`, after merging the
first two, the internal array is permuted to be
equivalent to `2 OUT HUP 01 HUP 1 OUT`).