Fix empty string case in parse_nonnegative_int
What's really annoying is that I feel like I used
to be better at catching these sorts of bugs when
initially implementing. And my old implementation
did not have these issues because of it. But here
I am, messing up now, many years later. Because I
have allowed my cognition to become sloppy in the
sense that it slips over the details to look at
the big picture.