C, like a gun, requires regular training
This is why in the past I got so opposed to `++`,
especially in combination with `*` dereferences.
It's so much easier to mentally move it around
while losing track of the order, than a `+= 1`
in a separate statement. When I first fixed this
I think I did `character = *string++`, but then
I moved the `++` down because it seemed more
fitting there, only to miss that it needed to
become a preincrement instead of a postincrement.

Anyway, I don't have a clear opinion yet on which
is better in this case - `*string++` in the loop
loop body or `*++string` in the `while` condition.