Fix events array iteration more
This is another good example of why postincrement is just
hostile to human reasoning - it takes more mental state
to remember that the increment has already happened when
pointer-walking, or more habitualized cognition to always
apply the postincrement at the right time rather than
flowing over it mentally with a more big-picture gloss
that conflates `while(foo++)` with `for(; foo; foo += 1)`.

What's annoying is that I've learned these lessons before,
and now I'm re-learning them because I've grown to like
`++` and `--` shorthand in certain cases, so now I'm all
over the place applying them, while being too attrophied
in the cognition needed to not logically slip with such
operators in code.