Stop accepting `--` for now.
It was a mistake in the sense of "I accidentally
decided to add support for this during the recent
refactoring, before remembering reasons against it."

`poll` takes no non-option arguments that can
validly start with a dash, so it never needs
a `--` to disambiguate anything.

It might be good to support `--` in all programs,
even the ones which never need `--`, so that we
can use the same patterns and habits with all of
the program invocations we do, but if so, then
the implementation this commit deletes is wrong,
because the universally consistent way is to also
allow `--` after any options, not only *instead*
of an option. (The "instead-of" way makes sense
in tools where there is only one option and it
has to be first, since that means only the
no-options case has ambiguity.)